# swagger/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "WHS-TodoList API",
    "version": "1.0",
    "description": "API specification for the WHS-TodoList application. Manages users, todos, and holidays.",
    "contact": {
      "name": "WHS-TodoList Team",
      "email": "contact@whs-todolist.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3001/api",
      "description": "Development server"
    },
    {
      "url": "https://your-domain.vercel.app/api",
      "description": "Production server"
    }
  ],
...
  "paths": {
...
    "/todos": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get list of todos",
        "operationId": "getTodos",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": ["active", "completed", "deleted"]
            }
          },
          { "name": "search", "in": "query", "schema": { "type": "string" } },
          {
            "name": "sortBy",
            "in": "query",
            "schema": { "type": "string", "enum": ["dueDate", "createdAt"] }
          },
          {
            "name": "order",
            "in": "query",
            "schema": { "type": "string", "enum": ["asc", "desc"] }
          }
        ],
        "responses": {
          "200": {
            "description": "A list of todos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "data": {
                      "type": "array",
                      "items": { "$ref": "#/components/schemas/Todo" }
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Todos"],
        "summary": "Create a new todo",
        "operationId": "createTodo",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateTodoRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Todo created successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" }
        }
      }
    },
    "/todos/{id}": {
      "get": {
        "tags": ["Todos"],
        "summary": "Get a single todo by ID",
        "operationId": "getTodoById",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Detailed information for a todo",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "put": {
        "tags": ["Todos"],
        "summary": "Update an existing todo",
        "operationId": "updateTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateTodoRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Todo updated successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      },
      "delete": {
        "tags": ["Todos"],
        "summary": "Move a todo to trash (soft delete)",
        "operationId": "deleteTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo moved to trash",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TodoSoftDeleteResponse"
                }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/todos/{id}/complete": {
      "patch": {
        "tags": ["Todos"],
        "summary": "Mark a todo as complete",
        "operationId": "completeTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo marked as complete",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoStatusResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
    "/todos/{id}/restore": {
      "patch": {
        "tags": ["Todos"],
        "summary": "Restore a todo from trash",
        "operationId": "restoreTodo",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "uuid" }
          }
        ],
        "responses": {
          "200": {
            "description": "Todo restored successfully",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/TodoRestoreResponse" }
              }
            }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": { "$ref": "#/components/responses/NotFound" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "Todo": {
        "type": "object",
        "properties": {
          "todoId": { "type": "string", "format": "uuid" },
          "userId": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "content": { "type": "string", "nullable": true },
          "startDate": { "type": "string", "format": "date", "nullable": true },
          "dueDate": { "type": "string", "format": "date", "nullable": true },
          "status": {
            "type": "string",
            "enum": ["active", "completed", "deleted"]
          },
          "isCompleted": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "deletedAt": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
...
      "CreateTodoRequest": {
        "type": "object",
        "required": ["title"],
        "properties": {
          "title": { "type": "string", "example": "프로젝트 마감" },
          "content": { "type": "string", "example": "PRD 작성 완료하기" },
          "startDate": {
            "type": "string",
            "format": "date",
            "example": "2025-11-25"
          },
          "dueDate": {
            "type": "string",
            "format": "date",
            "example": "2025-11-28"
          }
        }
      },
      "UpdateTodoRequest": {
        "type": "object",
        "properties": {
          "title": { "type": "string" },
          "content": { "type": "string", "nullable": true },
          "startDate": { "type": "string", "format": "date", "nullable": true },
          "dueDate": { "type": "string", "format": "date", "nullable": true }
        }
      },
...
      "TodoResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": { "$ref": "#/components/schemas/Todo" }
        }
      },
      "TodoStatusResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "data": {
            "type": "object",
            "properties": {
              "todoId": { "type": "string", "format": "uuid" },
              "status": { "type": "string", "example": "completed" },
              "isCompleted": { "type": "boolean", "example": true }
            }
          }
        }
      },
      "TodoSoftDeleteResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "todoId": { "type": "string", "format": "uuid" },
              "status": { "type": "string", "example": "deleted" },
              "deletedAt": { "type": "string", "format": "date-time" }
            }
          }
        }
      },
      "TodoRestoreResponse": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": true },
          "message": { "type": "string" },
          "data": {
            "type": "object",
            "properties": {
              "todoId": { "type": "string", "format": "uuid" },
              "status": { "type": "string", "example": "active" },
              "deletedAt": { "type": "object", "nullable": true }
            }
          }
        }
      },
...
      "Error": {
        "type": "object",
        "properties": {
          "success": { "type": "boolean", "example": false },
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string" },
              "message": { "type": "string" }
            }
          }
        }
      }
    },
    "responses": {
      "BadRequest": {
        "description": "Bad Request - Invalid input",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
      "Forbidden": {
        "description": "Forbidden - Insufficient permissions",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
      "NotFound": {
        "description": "Not Found - Resource not found",
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/Error" }
          }
        }
      },
...
    }
  }
}